Reorganized CompareDirectories examples to its own package
Also added ASTNodeUtilities
