+++ BEXCodeRefactoringExamples/src/main/java/info/codesaway/becr/comparedirectories/CompareDirectoriesOption.java
@@ -1,4 +1,4 @@
-package info.codesaway.becr.examples;
+package info.codesaway.becr.comparedirectories;
 
 public enum CompareDirectoriesOption {
 	EXCLUDE_DEFAULT_SUBSTITUTIONS
