+++ BEXCodeRefactoringExamples/src/main/java/info/codesaway/becr/comparedirectories/CompareJavaCodeInfo.java
@@ -1,4 +1,4 @@
-package info.codesaway.becr.examples;
+package info.codesaway.becr.comparedirectories;
 
 import java.util.List;
